Separate electron imports between renderer and main
