+++ src/common/safeIpc.ts
@@ -1,13 +1,4 @@
-import {
-    BrowserWindow,
-    IpcMainEvent,
-    IpcMainInvokeEvent,
-    IpcRendererEvent,
-    MessagePortMain,
-    WebContents,
-    ipcMain as unsafeIpcMain,
-    ipcRenderer as unsafeIpcRenderer,
-} from 'electron/common';
+import { type FileFilter, type OpenDialogReturnValue } from 'electron/common';
 import { MakeDirectoryOptions } from 'fs';
 import { Mode, ObjectEncodingOptions, OpenMode, PathLike } from 'original-fs';
 import { FileOpenResult, FileSaveResult, PythonInfo, Version } from './common-types';
@@ -26,14 +17,10 @@ export interface InvokeChannels {
     'get-backend-url': ChannelInfo<string>;
     'refresh-nodes': ChannelInfo<boolean>;
     'get-app-version': ChannelInfo<Version>;
-    'dir-select': ChannelInfo<Electron.OpenDialogReturnValue, [dirPath: string]>;
+    'dir-select': ChannelInfo<OpenDialogReturnValue, [dirPath: string]>;
     'file-select': ChannelInfo<
-        Electron.OpenDialogReturnValue,
-        [
-            filters: Electron.FileFilter[],
-            allowMultiple: boolean | undefined,
-            dirPath: string | undefined
-        ]
+        OpenDialogReturnValue,
+        [filters: FileFilter[], allowMultiple: boolean | undefined, dirPath: string | undefined]
     >;
 
     'file-save-json': ChannelInfo<void, [saveData: SaveData, savePath: string]>;
@@ -138,85 +125,3 @@ export type ChannelArgs<C extends keyof (InvokeChannels & SendChannels)> = (Invo
     SendChannels)[C]['args'];
 export type ChannelReturn<C extends keyof (InvokeChannels & SendChannels)> = (InvokeChannels &
     SendChannels)[C]['returnType'];
-
-interface SafeIpcMain extends Electron.IpcMain {
-    handle<C extends keyof InvokeChannels>(
-        channel: C,
-        listener: (
-            event: IpcMainInvokeEvent,
-            ...args: ChannelArgs<C>
-        ) => Promise<ChannelReturn<C>> | ChannelReturn<C>
-    ): void;
-    handleOnce<C extends keyof InvokeChannels>(
-        channel: C,
-        listener: (
-            event: IpcMainInvokeEvent,
-            ...args: ChannelArgs<C>
-        ) => Promise<ChannelReturn<C>> | ChannelReturn<C>
-    ): void;
-    on<C extends keyof SendChannels>(
-        channel: C,
-        listener: (event: IpcMainEvent, ...args: ChannelArgs<C>) => void
-    ): this;
-    once<C extends keyof SendChannels>(
-        channel: C,
-        listener: (event: IpcMainEvent, ...args: ChannelArgs<C>) => void
-    ): this;
-    removeAllListeners(channel?: keyof SendChannels): this;
-    removeHandler(channel: keyof InvokeChannels): void;
-    removeListener<C extends keyof SendChannels>(
-        channel: C,
-        listener: (event: IpcMainEvent | IpcMainInvokeEvent, ...args: ChannelArgs<C>) => void
-    ): this;
-}
-
-interface SafeIpcRenderer extends Electron.IpcRenderer {
-    invoke<C extends keyof InvokeChannels>(
-        channel: C,
-        ...args: ChannelArgs<C>
-    ): Promise<ChannelReturn<C>>;
-    on<C extends keyof SendChannels>(
-        channel: C,
-        listener: (event: IpcRendererEvent, ...args: ChannelArgs<C>) => void
-    ): this;
-    once<C extends keyof SendChannels>(
-        channel: C,
-        listener: (event: IpcRendererEvent, ...args: ChannelArgs<C>) => void
-    ): this;
-    postMessage(channel: keyof SendChannels, message: unknown, transfer?: MessagePort[]): void;
-    removeAllListeners(channel: keyof SendChannels): this;
-    removeListener<C extends keyof SendChannels>(
-        channel: C,
-        listener: (event: IpcRendererEvent, ...args: ChannelArgs<C>) => void
-    ): this;
-    send<C extends keyof SendChannels>(channel: C, ...args: ChannelArgs<C>): void;
-    sendSync<C extends keyof SendChannels>(channel: C, ...args: ChannelArgs<C>): void;
-    sendTo<C extends keyof SendChannels>(
-        webContentsId: number,
-        channel: C,
-        ...args: ChannelArgs<C>
-    ): void;
-    sendToHost<C extends keyof SendChannels>(channel: C, ...args: ChannelArgs<C>): void;
-}
-
-interface WebContentsWithSafeIcp extends WebContents {
-    invoke<C extends keyof SendChannels>(
-        channel: C,
-        ...args: ChannelArgs<C>
-    ): Promise<ChannelReturn<C>>;
-    postMessage(channel: keyof SendChannels, message: unknown, transfer?: MessagePortMain[]): void;
-    send<C extends keyof SendChannels>(channel: C, ...args: ChannelArgs<C>): void;
-    sendSync<C extends keyof SendChannels>(channel: C, ...args: ChannelArgs<C>): ChannelReturn<C>;
-    sendTo<C extends keyof SendChannels>(
-        webContentsId: number,
-        channel: C,
-        ...args: ChannelArgs<C>
-    ): void;
-    sendToHost<C extends keyof SendChannels>(channel: C, ...args: ChannelArgs<C>): void;
-}
-export interface BrowserWindowWithSafeIpc extends BrowserWindow {
-    webContents: WebContentsWithSafeIcp;
-}
-
-export const ipcMain = unsafeIpcMain as SafeIpcMain;
-export const ipcRenderer = unsafeIpcRenderer as SafeIpcRenderer;
