+++ src/main/gui/main-window.ts
@@ -7,7 +7,6 @@ import { BackendEventMap } from '../../common/Backend';
 import { Version } from '../../common/common-types';
 import { isMac } from '../../common/env';
 import { log } from '../../common/log';
-import { BrowserWindowWithSafeIpc, ipcMain } from '../../common/safeIpc';
 import { SaveFile, openSaveFile } from '../../common/SaveFile';
 import { ChainnerSettings } from '../../common/settings/settings';
 import { CriticalError } from '../../common/ui/error';
@@ -16,6 +15,7 @@ import { OpenArguments, parseArgs } from '../arguments';
 import { BackendProcess } from '../backend/process';
 import { setupBackend } from '../backend/setup';
 import { getRootDirSync } from '../platform';
+import { BrowserWindowWithSafeIpc, ipcMain } from '../safeIpc';
 import { writeSettings } from '../setting-storage';
 import { MenuData, setMainMenu } from './menu';
 import { addSplashScreen } from './splash';
