+++ src/main/gui/menu.ts
@@ -5,9 +5,9 @@ import os from 'os';
 import path from 'path';
 import { isMac } from '../../common/env';
 import { links } from '../../common/links';
-import { BrowserWindowWithSafeIpc } from '../../common/safeIpc';
 import { openSaveFile } from '../../common/SaveFile';
 import { getLogsFolder } from '../platform';
+import { BrowserWindowWithSafeIpc } from '../safeIpc';
 import { getCpuInfo, getGpuInfo } from '../systemInfo';
 
 export interface MenuData {
