+++ src/renderer/components/groups/NcnnFileInputsGroup.tsx
@@ -1,6 +1,6 @@
 import { memo, useEffect } from 'react';
 import { log } from '../../../common/log';
-import { ipcRenderer } from '../../../common/safeIpc';
+import { ipcRenderer } from '../../../renderer/safeIpc';
 import { getInputValue } from '../../../common/util';
 import { SchemaInput } from '../inputs/SchemaInput';
 import { GroupProps } from './props';
