+++ src/renderer/components/inputs/FileInput.tsx
@@ -16,7 +16,7 @@ import { useTranslation } from 'react-i18next';
 import { BsFileEarmarkPlus } from 'react-icons/bs';
 import { MdContentCopy, MdFolder } from 'react-icons/md';
 import { useContext } from 'use-context-selector';
-import { ipcRenderer } from '../../../common/safeIpc';
+import { ipcRenderer } from '../../../renderer/safeIpc';
 import { AlertBoxContext } from '../../contexts/AlertBoxContext';
 import { getSingleFileWithExtension } from '../../helpers/dataTransfer';
 import { useContextMenu } from '../../hooks/useContextMenu';
