+++ src/renderer/contexts/GlobalNodeState.tsx
@@ -28,7 +28,7 @@ import { log } from '../../common/log';
 import { getEffectivelyDisabledNodes } from '../../common/nodes/disabled';
 import { ChainLineage } from '../../common/nodes/lineage';
 import { TypeState } from '../../common/nodes/TypeState';
-import { ipcRenderer } from '../../common/safeIpc';
+import { ipcRenderer } from '../../renderer/safeIpc';
 import { ParsedSaveData, SaveData } from '../../common/SaveFile';
 
 import {
