+++ src/renderer/contexts/SettingsContext.tsx
@@ -2,7 +2,7 @@ import { useColorMode } from '@chakra-ui/react';
 import React, { SetStateAction, memo, useCallback, useEffect, useState } from 'react';
 import { createContext, useContext } from 'use-context-selector';
 import { log } from '../../common/log';
-import { ipcRenderer } from '../../common/safeIpc';
+import { ipcRenderer } from '../../renderer/safeIpc';
 import { ChainnerSettings, defaultSettings } from '../../common/settings/settings';
 import { noop } from '../../common/util';
 import { useMemoObject } from '../hooks/useMemo';
