+++ src/renderer/index.tsx
@@ -3,8 +3,8 @@ import path from 'path';
 import { createRoot } from 'react-dom/client';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { LEVEL_NAME, log } from '../common/log';
-import { ipcRenderer } from '../common/safeIpc';
 import { App } from './app';
+import { ipcRenderer } from './safeIpc';
 
 ipcRenderer
     .invoke('get-appdata')
