+++ src/renderer/splash.tsx
@@ -3,8 +3,8 @@ import { memo, useEffect, useState } from 'react';
 import { createRoot } from 'react-dom/client';
 import './i18n';
 import { useTranslation } from 'react-i18next';
-import { ipcRenderer } from '../common/safeIpc';
 import { ChaiNNerLogo } from './components/chaiNNerLogo';
+import { ipcRenderer } from './safeIpc';
 import { theme } from './splashTheme';
 
 const Splash = memo(() => {
